Lots of changes to make things work on Challenger given no site ip addresses to use for kh external network domain.  Made changes to allow kh external network to be nated and portforwarded by ionodes.
to do this we added the support to a ctl protocol on the kh link layer.  We use
this and and the special tree route 0 to inform all the nodes in a pset of
facts about their io node so that they can setup their gateway appropriately.

reworked khctl scripts to be cleaner site specific scripts rather than
an master script (was getting to complicated)